rembg-api/requirements.txt
# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Traitement d'image
rembg==2.0.56
Pillow==10.1.0
numpy==1.26.2

# Utilitaires
pydantic==2.5.0
python-dotenv==1.0.0

# Pour les tests (optionnel)
pytest==7.4.3
httpx==0.25.2